add authorization for pet creation & deletion

+++ src/actions/actions.ts
@@ -4,8 +4,9 @@ import prisma from "@/lib/db";
 import { revalidatePath } from "next/cache";
 import { sleep } from "@/lib/utils";
 import { petFormSchema, petIdSchema } from "@/lib/zodSchemas";
-import { signIn, signOut } from "@/lib/auth";
+import { auth, signIn, signOut } from "@/lib/auth";
 import bcrypt from 'bcryptjs';
+import { redirect } from "next/navigation";
 
 /*
     sleep function is used to simulate network delay
@@ -43,6 +44,11 @@ export async function LogOut() {
 export async function addPet(petData: unknown) {
     //  await sleep(2);
 
+    const session = await auth();
+    if (!session?.user) {
+        redirect("/login");
+    }
+
     const validatedPetData = petFormSchema.safeParse(petData);
     if (!validatedPetData.success) {
         return { message: "Invalid pet data was recieved by the server." }
@@ -50,7 +56,7 @@ export async function addPet(petData: unknown) {
 
     try {
         await prisma?.pet.create({
-            data: validatedPetData.data
+            data: { ...validatedPetData.data, user: { connect: { id: session.user.id } } }
         });
     } catch (error) {
         return { message: "There was a problem adding this pet." }
@@ -86,11 +92,30 @@ export async function editPet(petId: unknown, petData: unknown) {
 export async function deletePet(petId: unknown) {
     // await sleep(2);
 
+    // authentication check
+    const session = await auth();
+    if (!session?.user) {
+        redirect("/login");
+    }
+
+    // validation
     const validatedPetId = petIdSchema.safeParse(petId);
     if (!validatedPetId.success) {
         return { message: "No pet with this ID was found." }
     }
 
+    // authorization check
+    let requestedPet;
+    try {
+        requestedPet = await prisma.pet.findUnique({ where: { id: validatedPetId.data }, select: { userId: true } });
+    } catch (error) {
+        return { message: "There was a problem finding your pet." }
+    }
+    if (requestedPet!.userId !== session.user.id) {
+        return { message: "That's not your pet." }
+    }
+
+    // db action
     try {
         await prisma.pet.delete({ where: { id: validatedPetId.data } });
     } catch (error) {
